move 'hasIn' tests to separate file
