+++ types/ramda/tsconfig.json
@@ -89,6 +89,7 @@
         "test/groupWith-tests.ts",
         "test/gt-tests.ts",
         "test/gte-tests.ts",
-        "test/has-tests.ts"
+        "test/has-tests.ts",
+        "test/hasIn-tests.ts"
     ]
 }
